Replace all prompt pages with a single version
